requirements.txt
flask
flask-cors
transformers
torch
newspaper3k
pandas
numpy
requests
beautifulsoup4
lxml
pillow

pip freeze > requirements.txt
```

2. **Create `Procfile` (no extension) in project root:**
```
web: python src/api.py